♻️ Fold duplicated theme helpers into the shared modules
Heritage and practice research each carried a member paginator; practice
research also duplicated the viewer check and the license badge. The facet
guard moves onto ThemeHomepage so a new theme inherits it rather than copying it.
