+++ app/presenters/practice_research/homepage_presenter.rb
@@ -27,16 +27,6 @@ module PracticeResearch
 
     private
 
-    # A tenant can remove any facet from the catalog config, and Blacklight
-    # raises rather than returning nothing when asked to facet on a field it
-    # does not know, so a theme cannot assume a facet exists.
-    def facet_items(field, extra_params = {})
-      return [] unless @search_service.blacklight_config.facet_fields.key?(field)
-
-      @search_service.facet_field_response(field, { 'facet.mincount' => '1' }.merge(extra_params))
-                     .aggregations[field]&.items.to_a
-    end
-
     # The band holds four tiles on either axis. Solr sorts by index when the
     # limit is -1, so the largest are picked here rather than in the query.
     def model_counts
