+++ app/presenters/theme_homepage.rb
@@ -40,6 +40,16 @@ module ThemeHomepage
 
   private
 
+  # A tenant can remove any facet from the catalog config, and Blacklight
+  # raises rather than returning nothing when asked to facet on a field it
+  # does not know, so a theme cannot assume a facet exists.
+  def facet_items(field, extra_params = {})
+    return [] unless @search_service.blacklight_config.facet_fields.key?(field)
+
+    @search_service.facet_field_response(field, { 'facet.mincount' => '1' }.merge(extra_params))
+                   .aggregations[field]&.items.to_a
+  end
+
   # FeaturedWorkList and FeaturedCollectionList build their presenters with
   # ability = nil, so anything rendering them must re-check read access.
   def readable_ids(ids)
