+++ app/views/themes/heritage/hyrax/homepage/_share_your_work.html.erb
@@ -5,7 +5,7 @@
   </div>
   <div class="hrt-cta-actions">
     <%= link_to t(:'hyrax.pages.tabs.terms_page'), hyrax.terms_path, class: 'hrt-cta-terms' %>
-    <% href, data = signed_in? ? deposit_new_work_target(many: @presenter.create_many_work_types?, first_type: @presenter.first_work_type) : [hyrax.my_works_path, {}] %>
+    <% href, data = theme_deposit_target %>
     <%= link_to t('heritage.homepage.cta.button'), href, data:, class: 'hrt-cta-button' %>
   </div>
 </section>
