+++ app/views/themes/heritage_show/hyrax/base/_child_works_panel.html.erb
@@ -1,4 +1,4 @@
-<% paginated = hrt_child_work_ids(presenter) %>
+<% paginated = theme_child_work_ids(presenter) %>
 <% if paginated.any? %>
   <section class="hrt-panel" id="hrt-children" aria-labelledby="hrt-children-label">
     <div class="hrt-panel-head">
