+++ app/views/themes/heritage_show/hyrax/base/_file_panel.html.erb
@@ -1,4 +1,4 @@
-<% paginated = hrt_file_set_ids(presenter) %>
+<% paginated = theme_file_set_ids(presenter) %>
 <% if paginated.any? %>
   <section class="hrt-panel" id="hrt-files" aria-labelledby="hrt-files-label">
     <div class="hrt-panel-head">
