+++ app/views/themes/screening_room/hyrax/homepage/_share_your_work.html.erb
@@ -6,7 +6,7 @@
 
   <div class="scr-cta-actions">
     <%= link_to t(:'hyrax.pages.tabs.terms_page'), hyrax.terms_path, class: 'scr-cta-terms' %>
-    <% href, data = signed_in? ? deposit_new_work_target(many: @presenter.create_many_work_types?, first_type: @presenter.first_work_type) : [hyrax.my_works_path, {}] %>
+    <% href, data = theme_deposit_target %>
     <%= link_to t('screening_room.homepage.cta.button'), href, data:, class: 'scr-cta-button' %>
   </div>
 </section>
