+++ spec/helpers/practice_research_show_helper_spec.rb
@@ -55,34 +55,6 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
     expect(described_class::DEFAULT_ROWS).to eq(20)
   end
 
-  describe '#pr_viewer?' do
-    it 'is true for a work with a representative to show' do
-      presenter = double('presenter', video_embed_viewer?: false, representative_id: 'fs-1',
-                                      representative_presenter: double('file set'))
-
-      expect(helper.pr_viewer?(presenter)).to be(true)
-    end
-
-    it 'is true for an embed, which carries its own media' do
-      presenter = double('presenter', video_embed_viewer?: true)
-
-      expect(helper.pr_viewer?(presenter)).to be(true)
-    end
-
-    it 'is false with no representative, so the default work icon is not given a band' do
-      presenter = double('presenter', video_embed_viewer?: false, representative_id: nil)
-
-      expect(helper.pr_viewer?(presenter)).to be(false)
-    end
-
-    it 'is false when the representative id points at nothing indexed' do
-      presenter = double('presenter', video_embed_viewer?: false, representative_id: 'gone',
-                                      representative_presenter: nil)
-
-      expect(helper.pr_viewer?(presenter)).to be(false)
-    end
-  end
-
   describe 'member pane pagination' do
     let(:ids) { (1..12).map { |n| "work-#{n}" } }
     let(:presenter) { double('presenter', authorized_file_set_ids: %w[file-1], authorized_child_work_ids: ids) }
@@ -107,7 +79,7 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
     it 'keeps deposit order across the page boundary' do
       page_one = helper.pr_child_work_ids(presenter).to_a
 
-      helper.instance_variable_set(:@pr_child_work_ids, nil)
+      helper.instance_variable_set(:@theme_child_work_ids, nil)
       helper.params[:items_page] = '2'
 
       expect(page_one + helper.pr_child_work_ids(presenter).to_a).to eq(ids)
@@ -166,7 +138,7 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
     end
   end
 
-  describe '#pr_active_pane' do
+  describe '#theme_active_pane' do
     let(:presenter) do
       double('presenter', authorized_file_set_ids: Array.new(7) { |n| "f#{n}" },
                           authorized_child_work_ids: Array.new(12) { |n| "w#{n}" },
@@ -176,20 +148,20 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
     before { allow(helper).to receive(:pr_context_html).and_return('<p>Statement</p>') }
 
     it 'is the first pane with no page param' do
-      expect(helper.pr_active_pane(presenter)).to eq(:context)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:context)
     end
 
     it 'opens the pane the pager link names' do
       helper.params[:pane] = 'items'
 
-      expect(helper.pr_active_pane(presenter)).to eq(:items)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:items)
     end
 
     it 'opens it on page one too, where Kaminari drops the page param' do
       helper.params[:pane] = 'files'
 
       expect(helper.pr_file_set_ids(presenter).current_page).to eq(1)
-      expect(helper.pr_active_pane(presenter)).to eq(:files)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:files)
     end
 
     it 'ignores a pane that is not being shown' do
@@ -197,19 +169,19 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
                                            authorized_child_work_ids: Array.new(12) { |n| "w#{n}" })
       helper.params[:pane] = 'files'
 
-      expect(helper.pr_active_pane(presenter)).to eq(:context)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:context)
     end
 
     it 'ignores a pane that does not exist' do
       helper.params[:pane] = 'nonsense'
 
-      expect(helper.pr_active_pane(presenter)).to eq(:context)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:context)
     end
 
     it 'ignores an array pane instead of raising on it' do
       helper.params[:pane] = ['items']
 
-      expect(helper.pr_active_pane(presenter)).to eq(:context)
+      expect(helper.theme_active_pane(helper.pr_show_panes(presenter))).to eq(:context)
     end
   end
 
@@ -267,32 +239,6 @@ RSpec.describe PracticeResearchShowHelper, type: :helper do
 
     it 'drops the two the sidebar renders by hand, and any field with no value' do
       expect(helper.pr_card_fields(presenter)).to eq([:identifiers])
-    end
-  end
-
-  describe '#pr_license_badge' do
-    def badge_for(license)
-      helper.pr_license_badge(double('presenter', license: Array(license)))
-    end
-
-    it 'shortens a Creative Commons licence URL' do
-      expect(badge_for('https://creativecommons.org/licenses/by/4.0/')).to eq('CC BY 4.0')
-    end
-
-    it 'handles hyphenated codes' do
-      expect(badge_for('https://creativecommons.org/licenses/by-nc-nd/3.0/')).to eq('CC BY-NC-ND 3.0')
-    end
-
-    it 'recognises CC0' do
-      expect(badge_for('http://creativecommons.org/publicdomain/zero/1.0/')).to eq('CC0 1.0')
-    end
-
-    it 'is nil for anything else, so the badge is not rendered' do
-      expect(badge_for('http://www.europeana.eu/portal/rights/rr-r.html')).to be_nil
-    end
-
-    it 'is nil with no licence at all' do
-      expect(badge_for(nil)).to be_nil
     end
   end
 end
